Litte refactor v2. Moved find controllers to ClientConroller class.
