+++ PerfectWeddingWeb/src/main/resources/providers.json
@@ -380,4 +380,28 @@
   },
   "ratingList" : [ ],
   "active" : false
+}, {
+  "averageRating" : 0.0,
+  "currentID" : 17,
+  "companyName" : "test po refactor",
+  "ownerName" : "",
+  "ownerSurname" : "",
+  "phone" : "",
+  "email" : "",
+  "websiteAddress" : "",
+  "location" : {
+    "voivodeship" : "DOLNOŚLĄSKIE",
+    "city" : ""
+  },
+  "serviceType" : {
+    "description" : "",
+    "price" : "",
+    "typesOfService" : "FILMOWANIE",
+    "id" : 0
+  },
+  "availability" : {
+    "dates" : [ null ]
+  },
+  "ratingList" : [ ],
+  "active" : false
 } ]
